requirements.txt
langchain==0.1.17
langchain-community==0.0.36
langchain-weaviate==0.0.1.post1
weaviate-client==4.5.7
pypdf==4.2.0
streamlit==1.34.0
chardet==5.2.0